FIBER_TASKING_LIB: Use a std::vector protected by a CRITICAL_SECTION for holding waiting tasks
Instead of completely emptying the a lockless queue during search and then pushing the not-ready
tasks back into the lockless queue. Not only was this extremely inefficient, the temporary there is a
chance that the memory allocated by the temporary vector would never be destructed, since we could
switch fibers and never come back.

The problem is the following:
- We need to iterate through the list to check if there are any waiting tasks that we can switch to.
- If we find one, we need to remove it form the list, and then switch to it

My first instinct was some kind of read-write lock. However, this ends up failing, because it's likely that
two thread will end up finding the same task. As well as it ends up being just repetitive work. So, yes,
all the cores are active by allowing reads to happen at the same time. But they're doing the same thing.
So, we might as well let some of the cores sleep and do something else.

So in the end, it's just a single mutex that protects the entire process.
- Lock
- Loop through the list to find the first ready task.
    - If we find one,
        - Remove it from the list
        - break
- Unlock
- If we found one
    - Switch fibers

We don't switch fibers while in the loop because we have to unlock the mutex first