FIBER_TASKING_LIB: Create a switching fiber for each thread
In order to put the current fiber on the waitingTasks list or the fiber pool, we have to
switch to a different fiber. Otherwise, another thread could pop from the list/pool and
try to execute the current fiber before we have time to switch.

So, we temporarily store the relavent info in thread local storage and switch to a helper fiber.
Then the helper fiber stores everything properly and does the fiber swap.

The problem was that every so often, two threads would want to swap at the same time.
If there is only one helper fiber, the threads will try to simultaneously start executing the same
fiber. Which causes all kinds of havok and undefined behavior. (Which took FOREVER to
try to debug). To fix the problem, we create one helper fiber per thread.