FIBER_TASKING_LIB: Correct thread cleanup during Quit()
Firstly, there was an error in the if() compare. Once that was fixed, I realized that
the a thread could 'finish' while in a different execution context (aka, running a fiber
that wasn't its 'original' fiber). If this happens, calling FTLThreadJoin() on the original
ThreadIds will never return. Therefore, we have to manually track when threads finish,
(this time by tracking when a fiber breaks out of the infinite while loop), then manually
clean up the threads, since the threads will never get the chance to clean themselves up