FIBER_TASKING_LIB: Create a config file to define what platforms can support
This makes it easier for the end user to test specifically what they're after, rather
than having to remember which platforms support which features. It will also make
maintenance easier in the future. ie: to add a new platform (or buildchain for an existing
platform), we just modify config.h

Closes #4