FIBER_TASKING_LIB: Implement a function to get a unique integral id for a thread
Previously, 'fake' tls was using FTLGetCurrentThread() and incorrectly assuming the
return is an integral unique number. On Windows, it is integral (it's a HANDLE), but in
this case, the HANDLE is just a redirection handle. ie: it's value is identical for every
thread. Thankfully, Win32 has GetCurrentThreadId(), so no big problem. Posix, on the
other hand has no function to get a unique id for the thread. So we have to turn to OS
specific methods. In this case, gettid.

This commit should correct the errors thrown by the TLS unit tests. Issue #9