FIBER_TASKING_LIB: Portably detect std:align implementation
GCC doesn't have std::align until version 5.1. That said, std::align isn't
implemented by GCC, but by the std lib, ie. libstdc++ 3. On linux, clang
uses libstdc++ as well, but not necessarily the same version. So, in order
to test for std::align, we have to test what type and version of the std
lib we have. We borrow some code from Boost.Config to detect the type and
version.