Clean up the old fiber before adding a waiting fiber
Ok, so it turns out my massive paragraph of "Explanation", where I
thought I was soooo clever with my proof and QED, had a flaw.

While it *is* correct that all code paths will end up at the top of the
while loop in FiberStart, WaitForCounter writes to OldFiberIndex and
OldFiberDestination. So, we have to clean up the old fiber in
WaitForCounter in addition to at the top of FiberStart.

This fixes fibers "leaking" from the fiber pool