Initialize m_threads in the thread itself, rather than in main
When we create a thread it starts immediately. We set m_threads[i] right
after creating, but there is a delay. So if the worker thread starts
very fast, it can execute code that relies on m_threads before m_threads
is set.

To prevent this, we just set m_htreads[index] in the thread itself. In
this way, we guarantee that a thread has an entry before it tries to use
it.