Use a compiler barrier instead of a memory barrier on strong memory model CPU architectures
As pointed out in Issue #27, the memory synchronization in Steal() is a
large portion of the overhead in the library. If we detect the CPU
architecture, we can optimize this slightly. Strong memory model CPUs
will guarantee that, at the high level, memory stores and reads will
happen in the order of the instructions. That said, we still need to use
a compiler barrier to ensure it doesn't try to re-order reads and stores.

Even with this change, memory synchronization is a large portion of the
overhead. More work can be done.