First-try implementation of Event-based Waitlist
Instead of polling the list of potential waiting fibers every time we
try to find a new task, we turn the problem on its head. We wrap the
atomic counters, so we can track when they change. Then, we can identify
when a waiting fiber is ready.

Comments:
- This current implementation does improve speed, however, there is a
lot of contention on the locks in AtomicCounter. The old implementation
has no contention, but has to fetch a lot more memory and do more work.
- For 99% of the counter value changes, there *will* be a fiber waiting
on a counter, but it will *not* be ready. It would be nice if we could
have a fast path for the 99%, and a slow path for when there is a fiber
ready.
- Ready/Write lock isn't going to help any, since memory writes on one
core are going to force a cache invalidation on the other cores.