Fix a race condition / wrong logic in AtomicCounter
In 328303061b98e5f4b9a7172eb65bef67781dde4e, I added code that was
supposed to do a final check after adding a fiber to the wait list. However,
the implementation was completely broken.

First, we called continue instead of return when we failed the race, which
would cause us to store the waiting fiber *again*

Second, AddFiberToWaitingList didn't signal the TaskScheduler caller if
we hit the final value and finished. So, the TaskScheduler blindly switched
to a new fiber and waited on a counter that would never trigger.

Lastly, I changed the memory_order of some of the loads, in order to match
how they are being stored