Fix hang caused by AtomicCounter
memory_order_release guarantees that stores above the line will be visible to all
threads after they use memory_order_acquire. That said, it is perfectly valid for the
compiler / memory model to move loads that are below the line to above the line.

This causes our values to be stale. If we use memory_order_seq_cst, it will force
an implicit memory barrier, keeping the two areas separate.