Add the beginning implementation of empty queue behavior
The idea is to be able to control how much CPU the worker threads are consuming.
In Spin mode, FTL assumes that, for the majority of the time, there is work in the 
queues. Therefore, the worker threads will continually spin, searching for more 
Tasks to run. This allows us to have relatively low latency between Tasks being 
added to the queues and the workers starting work.

However, if the workload is lighter, the worker threads will blindly consume 100% 
CPU doing useless work searching for Tasks that aren't there. Thus, the two other 
behaviors aim to help mitigate consuming all of the CPU.

The easiest / simplest way to gain back some CPU is to just yield any remaining CPU 
time back to the OS once we've searched all the queues for tasks. This is exactly 
what the Yield behavior does.

The Sleep behavior will be more complicated. We need to strike the balance 
between saving CPU time and the overhead and latency of sleeping / waking a 
thread.