Merge the pinned and non-pinned fiber code paths
martty found a pretty big race condition between the pinned and non-pinned
code paths. In the non-pinned code path, waiting fibers are stored inside 
their relevant AtomicCounter. Whenever the counter is modified, we check
the waiting fibers to see if any of their conditions are fulfilled. If so, we add
it to our own thread's "ReadyFibers" list, to be taken in the next loop.

Pinned fibers are added to a TLS PinnedFibers list, which is polled every loop,
to check if the counter equals the correct value.

The problem arises as follows: An AtomicCounter decrements. After the
atomic decrement, it calls interal code to check if any fibers are waiting
on this new value. However, between the decrement and the internal
code, another thread checks its "PinnedFibers" list, finds the fiber's
counter to be correct value, and switches to it. In martty's case, the
counter is allocated on the stack, and returning from the wait causes
the scope to change and the counter to deallocate.

Back in the AtomicCounter, it's still trying to call internal code. As soon as
it references a member variable, it SegFaults, since the AtomicCounter
has been deallocated.

In order to solve this, I've merged the two code paths. Pinned fibers are
tracked in the AtomicCounters just like non-pinned. The only difference
is what "ReadyFibers" list they get added to when they're ready.

In the original code, ready fibers were always added to the current
thread's list, because they we could ignore sychronization and locks.
Now that we're having multithreaded access to the ReadFibers lists,
I added simple SpinLocks to each list. The locks should have very 
little contention, since the number of pinned tasks should be relatively
low in comparison to non-pinned. Non-pinned tasks will get added
to the current thread's list, like before. So this case has no contention