Fix race in fiber readying
When a counter is allocated on the stack, it is imperative that the fiber owning the counter does not resume before the other fibers finish, as this would destruct the counter while the other threads interact with it. This fix adds an atomic counter to ensure all threads finish accessing the counter before allowing fiber readying. Note that this does not help with concurrent readying, but in that case the counter needs to have it's lifetime externally managed anyway.