Kick sleeping threads when resuming Tasks that are pinned
See Issue #43

For example, the worker threads start and immediately go to sleep. 
When the main task adds tasks, the workers correctly wake up and start 
churning through the tasks. The main task is swapped out, so the main
thread also helps work on tasks.

When the counter is decremented to zero, that thread will detect that 
the "end" condition is met. If the Task isn't pinned, that thread will add the 
Task to its own queue to resume. If the Task is pinned, the thread adds the 
Task to the correct thread's queue.

The problem is that the other thread could be asleep.