Set the thread Handle for the main thread to INVALID_HANDLE_VALUE
::GetCurrentThread is a pseudo handle, that always references the current 
thread. Aka, if we tried to use this handle from another thread to reference
the main thread, it would instead reference the other thread. We don't 
currently use the handle anywhere. Therefore, we set this to 
INVALID_HANDLE_VALUE, so any future usages can take this into account

Thanks @rtj for reporting this. :)