Use an atomic bitfield for m_freeFibers instead of an array of atomic<bool>
The idea being that we can use less atomic memory fetches / memory
thrashing.

Another approach would be to still use the array of atomic<bool>,
but have the threads start from a random index each time. To do
this we would need to have a prng initialized and seeded in TLS.
(PCG could work well / be fast and simple).

Thus, GetNextFreeFiberIndex() would first query the prng in order to
get the start index, and then it would search from startIndex and loop
around to startIndex - 1. (the current behavior iterates from 0 to
m_fiberPoolSize)