WIP More progress on getting WASM working
It compiles fine now. And individual tests work.
But if I try to enable multiple tests together, it will hang.
It happens during the TaskScheduler deconstructor.

The worker threads are failing to switch to the fiber
from the pool. It never arrives. So the shutdown hangs,
because the threads are stuck in limbo.

My current thought is we're leaking something. Maybe
we're not doing something correct with posix threads.
Or we need to clean up the fibers. Or something else?